+++ src/main/java/kitchenpos/product/infrastructure/JpaProductRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu.infrastructure.persistence;
+package kitchenpos.product.infrastructure;
 
-import kitchenpos.menu.domain.model.Product;
-import kitchenpos.menu.domain.repository.ProductRepository;
+import kitchenpos.product.domain.model.Product;
+import kitchenpos.product.domain.repository.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
