+++ src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu.ui;
+package kitchenpos.product.ui;
 
-import kitchenpos.menu.application.ProductService;
-import kitchenpos.menu.domain.model.Product;
+import kitchenpos.product.application.ProductService;
+import kitchenpos.product.domain.model.Product;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
