+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -5,12 +5,12 @@ import kitchenpos.menu.domain.model.Menu;
 import kitchenpos.menu.domain.repository.MenuGroupRepository;
 import kitchenpos.menu.domain.model.MenuProduct;
 import kitchenpos.menu.domain.repository.MenuRepository;
-import kitchenpos.menu.domain.model.Product;
-import kitchenpos.menu.domain.repository.ProductRepository;
+import kitchenpos.product.domain.model.Product;
+import kitchenpos.product.domain.repository.ProductRepository;
 import kitchenpos.menu.domain.service.PurgomalumClient;
 import kitchenpos.menu.infrastructure.persistence.InMemoryMenuGroupRepository;
 import kitchenpos.menu.infrastructure.persistence.InMemoryMenuRepository;
-import kitchenpos.menu.infrastructure.persistence.InMemoryProductRepository;
+import kitchenpos.product.infrastructure.InMemoryProductRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
