+++ src/test/java/kitchenpos/product/infrastructure/InMemoryProductRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu.infrastructure.persistence;
+package kitchenpos.product.infrastructure;
 
-import kitchenpos.menu.domain.model.Product;
-import kitchenpos.menu.domain.repository.ProductRepository;
+import kitchenpos.product.domain.model.Product;
+import kitchenpos.product.domain.repository.ProductRepository;
 
 import java.util.ArrayList;
 import java.util.HashMap;
